loglevel default ctor takes bool value.
The loglevel default ctor takes a boolean value to indicate its inital
log level. This can be used to create const loglevels that can be
deactivated when created.

A copy ctor was also added that takes a boolean value, in order to
copy but deactive the copied log independent of the original.